Module Lsp.Cli

Handling of standard lsp server command line arguments

module Channel : sig ... end

The channel the server shold use to listen for connections

module Arg : sig ... end

Parsing of the standard commnad line arguments using Stdlib.Arg

val args : ?channel:Channel.t -> ?clientProcessId:int -> unit -> string list

generate command line arguments that can be used to spawn an lsp client