Module Cli.Arg

Parsing of the standard commnad line arguments using Stdlib.Arg

type t

val create : unit -> t

create () create a new record for arguments

val spec : t -> (string * Stdlib.Arg.spec * string) list

spec t returns the spec that should be provided to Stdlib.Arg to populate t using the interpreted cli args

val channel : t -> (Channel.t, string) result

channel t return the channel if correctly supplied. An error if the arguments were provided incorrectly.

val clientProcessId : t -> int option

Return the process id of the client used to run the lsp server if it was provided